Added fullness checks to ADTs
EAch ADT can now check if they have maxed out their size. This is useful
if you are using an array-based implementation and want to limit memory
use. Unfortunately there are no exceptions implemented yet, so those
should be added to take advantage of this.

Also made tiny tweaks to the build file to reduce fluff.