Added tests, fixed BinaryTree's toStrings
Simple test cases for the linear data structures all pass. In doing so,
found+fixed a bug in the DLL ADT.

Started on fixing up the Binary Tree ADT. Currently only fixed the
toString implementation, but not satisfied with the requirement of an
external variable. Might try reimplementing as level-order instead of
in-order?